modify how errors are created; add composer and remove initialize
