+++ lib/json_schematize/errors.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module JsonSchematize
+  class Error < StandardError; end
+  class ConfigError < StandardError; end
+  class FieldError < Error; end
+  class InvalidField < Error; end
+  class InvalidFieldByValidator < InvalidField; end
+  class InvalidFieldByType < InvalidField; end
+  class InvalidFieldByArrayOfTypes < InvalidField; end
+end
